refactored utils into NFACT_base
