+++ NFACT/NFACT_GLM/nfact_glm_args.py
@@ -1,5 +1,5 @@
 import argparse
-from NFACT.utils.utils import colours
+from NFACT.NFACT_base.utils import colours
 
 
 def nfact_glm_args() -> dict:
