+++ NFACT/NFACT_GLM/nfact_glm_setup.py
@@ -1,4 +1,5 @@
-from NFACT.NFACT_decomp.utils.utils import error_and_exit
+from NFACT.NFACT_base.utils import error_and_exit
+
 
 def check_complusory_args(args: dict) -> None:
     complusory_args = []
