+++ NFACT/NFACT_decomp/setup/configure_setup.py
@@ -1,10 +1,6 @@
 import os
-from NFACT.utils.utils import (
-    error_and_exit,
-    read_file_to_list,
-    colours,
-    load_json,
-)
+from NFACT.NFACT_base.utils import error_and_exit, colours
+from NFACT.NFACT_base.filesystem import read_file_to_list, load_json
 from NFACT.NFACT_decomp.pipes.image_handling import (
     check_files_are_imaging_files,
     get_imaging_details_from_path,
